Move ServerRenderer into their own file.
