+++ examples/simple_ssr/src/main.rs
@@ -95,7 +95,7 @@ async fn render() -> String {
         let rt = Builder::new_current_thread().enable_all().build().unwrap();
 
         set.block_on(&rt, async {
-            let renderer = yew::YewServerRenderer::<App>::new();
+            let renderer = yew::ServerRenderer::<App>::new();
 
             renderer.render().await
         })
